+++ src/main/java/site/siredvin/progressiveperipherals/common/tileentities/base/MutableNBTTileEntityEmbed.java
@@ -1,9 +1,10 @@
 package site.siredvin.progressiveperipherals.common.tileentities.base;
 
 import net.minecraft.block.BlockState;
+import net.minecraft.client.Minecraft;
 import net.minecraft.tileentity.TileEntity;
+import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
-import net.minecraftforge.client.model.ModelDataManager;
 import net.minecraftforge.common.util.Constants;
 
 public class MutableNBTTileEntityEmbed {
@@ -12,13 +13,18 @@ public class MutableNBTTileEntityEmbed {
         World world = tileEntity.getLevel();
         if (world != null) {
             if (world.isClientSide) {
-                ModelDataManager.requestModelDataRefresh(tileEntity);
+                tileEntity.requestModelDataUpdate();
+                BlockPos pos = tileEntity.getBlockPos();
+                // Basically, just world.setBlocksDirty with bypass model block state check
+                Minecraft.getInstance().levelRenderer.setBlocksDirty(
+                        pos.getX(), pos.getY(), pos.getZ(), pos.getX(), pos.getY(), pos.getZ()
+                );
             } else {
                 tileEntity.setChanged();
                 world.setBlockAndUpdate(tileEntity.getBlockPos(), state);
                 world.sendBlockUpdated(
                         tileEntity.getBlockPos(), tileEntity.getBlockState(), tileEntity.getBlockState(),
-                        Constants.BlockFlags.BLOCK_UPDATE + Constants.BlockFlags.NOTIFY_NEIGHBORS
+                        Constants.BlockFlags.BLOCK_UPDATE | Constants.BlockFlags.NOTIFY_NEIGHBORS
                 );
             }
         }